#2490 Condenser fan anti sweat behavior (#2500)
* add anti sweat behavior vote for condenser fan speed

* update type

* fix fan speed

* fix wrong ERD type

* update ERD lock

* update

* add swap

* update GUI

* alignment

* address comment

* address comment

* update

* update test cases and code to calculate cooling mode fan speed

* Initial condenser fan anti sweat behavior

* Revert "Initial condenser fan anti sweat behavior"

This reverts commit c329192ff957528bf9a09e883dd989b366f2219b.

* address comment

* add a test to check fan speed off

* update more

* update

* update test name

* add more tests

+++ Parametric/spec/Defrost/Defrost_spec.lua
@@ -757,12 +757,12 @@ describe('Defrost', function()
           u8(20),
           u8(]] .. sealed_system_valve_position_type.position_B .. [[),
           u8(]] .. compressor_speed_type.low_speed .. [[),
-          u8(]] .. fan_speed_type.super_low_speed .. [[),
-          u8(]] .. fan_speed_type.low_speed .. [[),
-          u8(]] .. fan_speed_type.medium_speed .. [[),
-          u8(]] .. fan_speed_type.high_speed .. [[),
+          u16(]] .. fan_speed_type.super_low_speed .. [[),
+          u16(]] .. fan_speed_type.low_speed .. [[),
+          u16(]] .. fan_speed_type.medium_speed .. [[),
+          u16(]] .. fan_speed_type.high_speed .. [[),
           u8(]] .. damper_position_type.open .. [[),
-          u8(]] .. fan_speed_type.super_low_speed .. [[),
+          u16(]] .. fan_speed_type.super_low_speed .. [[),
           u8(]] .. damper_position_type.closed .. [[),
           i16(-600),
           i16(3200),
@@ -806,7 +806,7 @@ describe('Defrost', function()
           i16(-1200),
           u8(]] .. sealed_system_valve_position_type.position_B .. [[),
           u8(]] .. compressor_speed_type.low_speed .. [[),
-          u8(]] .. fan_speed_type.low_speed .. [[),
+          u16(]] .. fan_speed_type.low_speed .. [[),
           u8(]] .. damper_position_type.closed .. [[)
         )
       )
